+++ ledgereth/__init__.py
@@ -1,6 +1,7 @@
 # flake8:noqa
 from ledgereth.objects import Transaction, SignedTransaction
-from ledgereth.web3 import get_account_by_path, get_accounts, sign_transaction, create_transaction
+from ledgereth.accounts import get_account_by_path, get_accounts
+from ledgereth.transactions import sign_transaction, create_transaction
 
 __version__ = '0.1.1'
 __author__ = 'Mike Shultz'
